# 4 commit - teste - Mudancas

+++ src/main/java/com/tuyo/funcionariosdata/repository/FuncionarioRepository.java
@@ -1,7 +1,7 @@
 package com.tuyo.funcionariosdata.repository;
 
-import com.tuyo.funcionariosdata.funcionario.Funcionario;
+import com.tuyo.funcionariosdata.entities.Funcionario;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FuncionarioRepository extends CrudRepository<Funcionario, Long> {
+public interface FuncionarioRepository extends CrudRepository<Funcionario, Integer> {
 }
